@@ -1,13 +1,13 @@
 <?php
 
-namespace S4mpp\Format;
+namespace S4mpp\Format\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
 /**
  * @see \S4mpp\Format\Skeleton\SkeletonClass
  */
-class FormatFacade extends Facade
+class Format extends Facade
 {
     /**
      * Get the registered name of the component.
